+++ angular-architecture/src/app/app.module.ts
@@ -2,17 +2,27 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { HttpModule } from '@angular/http';
+import { RouterModule } from '@angular/router';
 
 import { AppComponent } from './app.component';
+
+ // application modules 
+import { CommercialModule } from './commercial/commercial.module';
+import { CoreModule } from "./core/core.module";
 
 @NgModule({
   declarations: [
     AppComponent
   ],
   imports: [
+    // angular modules
     BrowserModule,
     FormsModule,
-    HttpModule
+    HttpModule,
+    RouterModule,
+    // application modules
+    CommercialModule,
+    CoreModule
   ],
   providers: [],
   bootstrap: [AppComponent]