+++ angular-architecture/src/app/commercial/commercial.module.ts
@@ -1,0 +1,13 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { CommercialRoutingModule } from './commercial-routing.module';
+
+@NgModule({
+  imports: [
+    CommonModule,
+    CommercialRoutingModule
+  ],
+  declarations: []
+})
+export class CommercialModule { }