+++ angular-architecture/src/app/commercial/products/products-routing.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+
+import { ProductsListComponent } from './products-list/products-list.component';
+import { ProductsFormComponent } from './products-form/products-form.component';
+
+const routes: Routes = [
+  { path: 'new', component: ProductsFormComponent },
+  { path: ':id', component: ProductsFormComponent },
+  { path: '', component: ProductsListComponent },
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule]
+})
+export class ProductsRoutingModule { }