+++ angular-architecture/src/app/commercial/products/products.module.ts
@@ -1,0 +1,18 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { ProductsRoutingModule } from './products-routing.module';
+import { ProductsListComponent } from './products-list/products-list.component';
+import { ProductsFormComponent } from './products-form/products-form.component';
+
+@NgModule({
+  imports: [
+    CommonModule,
+    ProductsRoutingModule
+  ],
+  exports:[
+    ProductsRoutingModule
+  ],
+  declarations: [ProductsListComponent, ProductsFormComponent]
+})
+export class ProductsModule { }