+++ angular-architecture/src/app/core/internal-error/internal-error.module.ts
@@ -1,0 +1,14 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { InternalErrorRoutingModule } from './internal-error-routing.module';
+import { InternalErrorPageComponent } from './internal-error-page/internal-error-page.component';
+
+@NgModule({
+  imports: [
+    CommonModule,
+    InternalErrorRoutingModule
+  ],
+  declarations: [InternalErrorPageComponent]
+})
+export class InternalErrorModule { }