+++ CHANGELOG.md
@@ -1,9 +1,22 @@
-# 0.2.0
+# CHANGELOG
+
+## 0.4.0
+
+- Adição do suporte à `Lazy Loading`
+- Criação dos módulos `commercial/products`, `core/dashboard`, `core/not-found`, `core/internal-error`
+- Criação dos componentes de layout `core/page-layout` e `core/panel-layout`
+
+## 0.3.0
+
+- Adição do suporte à rotas
+- Criação dos módulos `commercial` e `core` como exemplos para a utilização do suporte à rotas
+
+## 0.2.0
 
 - Adição do novo ambiente de desenvolvimento `Homologação (environment.stage.ts)`
 - Adição e suporte à testes com diferentes ambientes `Devenvolvimento (dev)`, `Homologação (stage)` e `Produção (prod)`
 
-# 0.1.0
+## 0.1.0
 
 - Criação do template básico utilizando `Angular CLI`
 - Adição do suporte a estilos `SCSS`